Automatic Removal of Subtasks
Analogous to automatic task/widget update upon subtask removal.

However, if a subtask is removed from a task, and if the subtask's TaskWidget is open, it is possible to  further modify it without seeing any update on the main task. This is because, upon subtask removal from the main task, their connection is also removed (subtask.parent is None). Therefore, the subtask effectively becomes a top-level task.